Switch between VRControls and OrbitControls
Switch between OrbitControls and VRControls when VR is enabled and otherwise. Otherwise touch events are not handled on mobile.

Also, added OrbitControls to the third party directory since we need to change the touchmove events to not be passive in order to call preventdefault on them as per (https://www.chromestatus.com/features/5093566007214080)